Add usePrefetching composition
